Create blank.yml
Wanna we want to try a new way to change this ethical World by using new ways of changing it.
We should commit something new that we make our world easy. This is full of adventure and ethics if you have any ideas about it can access to my email shuklajatin5555@gmail.com it is probably about 7 to 10 days project i.e., We want  to hack WhatsApp Messenger/business that is mostly use by Indian and further country that makees us a new delight that helpful in changing the way of thinking about hacking.